Move InstanceCreator to Bootstrap/Instances.
