+++ Source/Bifrost.Specs/Bifrost.Specs.csproj
@@ -212,12 +212,12 @@
     <Compile Include="Concepts\for_Value\when_getting_hashcode_for_values_with_transposed_values.cs" />
     <Compile Include="Concepts\given\concepts.cs" />
     <Compile Include="Bootstrap\Assemblies\for_AssemblyProvider\when_loading_an_assembly_triggers_another_assembly_load.cs" />
-    <Compile Include="Execution\for_InstanceCreator\given\an_instance_creator.cs" />
-    <Compile Include="Execution\for_InstanceCreator\when_implementing_type_does_not_have_default_constructor.cs" />
-    <Compile Include="Execution\for_InstanceCreator\when_multiple_implementing_types_exist.cs" />
-    <Compile Include="Execution\for_InstanceCreator\when_single_implementing_type_exist.cs" />
-    <Compile Include="Execution\for_InstanceCreator\when_no_implementing_types_exist.cs" />
-    <Compile Include="Execution\for_InstanceCreator\ITestInterface.cs" />
+    <Compile Include="Bootstrap\Instances\for_InstanceCreator\given\an_instance_creator.cs" />
+    <Compile Include="Bootstrap\Instances\for_InstanceCreator\when_implementing_type_does_not_have_default_constructor.cs" />
+    <Compile Include="Bootstrap\Instances\for_InstanceCreator\when_multiple_implementing_types_exist.cs" />
+    <Compile Include="Bootstrap\Instances\for_InstanceCreator\when_single_implementing_type_exist.cs" />
+    <Compile Include="Bootstrap\Instances\for_InstanceCreator\when_no_implementing_types_exist.cs" />
+    <Compile Include="Bootstrap\Instances\for_InstanceCreator\ITestInterface.cs" />
     <Compile Include="Configuration\for_ConfigurationExtensions\when_initializing_multiple_bindings_for_specific_storage.cs" />
     <Compile Include="Configuration\for_ConfigurationExtensions\SomeOtherType.cs" />
     <Compile Include="Configuration\for_ConfigurationExtensions\SomeType.cs" />
