+++ Source/Bifrost.Specs/Bootstrap/Instances/for_InstanceCreator/ITestInterface.cs
@@ -1,0 +1,6 @@
+﻿namespace Bifrost.Specs.Bootstrap.Instances.for_InstanceCreator
+{
+    public interface ITestInterface
+    {
+    }
+}
