--- Source/Bifrost.Specs/Execution/for_InstanceCreator/ITestInterface.cs
@@ -1,6 +1,0 @@
-﻿namespace Bifrost.Specs.Execution.for_InstanceCreator
-{
-    public interface ITestInterface
-    {
-    }
-}
