+++ Source/Bifrost/Bifrost.csproj
@@ -124,10 +124,10 @@
     <Compile Include="Concepts\Value.cs" />
     <Compile Include="Concurrency\IScheduler.cs" />
     <Compile Include="Concurrency\Scheduler.cs" />
-    <Compile Include="Execution\IInstanceCreator.cs" />
-    <Compile Include="Execution\InstanceCreator.cs" />
     <Compile Include="Configuration\Assemblies\AssemblyConfigurationExtensions.cs" />
     <Compile Include="Configuration\Assemblies\AssembliesStartingWith.cs" />
+    <Compile Include="Bootstrap\Instances\IInstanceCreator.cs" />
+    <Compile Include="Bootstrap\Instances\InstanceCreator.cs" />
     <Compile Include="Configuration\Assemblies\IAssembliesConfiguration.cs" />
     <Compile Include="Bootstrap\Assemblies\IncludeNone.cs" />
     <Compile Include="Specifications\MatchAll.cs" />
